Created Procedure
This is a Kaseya Procedure designed to mass update Veeam instances. The creation of this was complicated by having to uncover several undocumented MSI flags that were required, found via SuperOrca MSI editor. Furthermore the documentation provided by Veeam at the time did not mention many of these steps, which were necessary to function properly. Note this is designed as an in place upgrade from any previous 9.5 build.